+++ change_log.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 - Hide borders on header and footer components
 - Updated the data, content for the actual resume
+- Add button to scroll to the next section in footer component
 
 ## 2025-04-25